test(ActionDecoderBase): enhanced EditMode coverage
