+++ unity-ggjj/Assets/Scripts/TextDecoder/Parser/MethodNotFoundScriptParsingException.cs
@@ -2,7 +2,7 @@ namespace TextDecoder.Parser
 {
     public class MethodNotFoundScriptParsingException : ScriptParsingException
     {
-        public MethodNotFoundScriptParsingException(string className, string methodName) : base($"Class '{className}' contains no method '{methodName}()'")
+        public MethodNotFoundScriptParsingException(string className, string methodName) : base($"Class '{className}' contains no non-public method '{methodName}()'")
         {
         }
     }
