+++ unity-ggjj/Assets/Tests/EditModeTests/TextDecoder/ActionDecoderTests/UnsupportedParameterTypeTests.cs
@@ -1,0 +1,28 @@
+using NUnit.Framework;
+using TextDecoder.Parser;
+using UnityEngine;
+
+namespace TextDecoder.ActionDecoderTests
+{
+    public class UnsupportedParameterTypeTests
+    {
+        private class ActionDecoderExtension : ActionDecoderBase
+        {
+            protected void NOT_SUPPORTED_PARAMETER(object unsupported)
+            {
+                // intentionally void
+            }
+        }
+
+        [Test]
+        public void ThrowsMissingParserException()
+        {
+            const string LINE_TO_PARSE = "&NOT_SUPPORTED_PARAMETER:cantBeParsed";
+
+            Debug.Log("Attempting to parse:\n" + LINE_TO_PARSE);
+            Assert.Throws<MissingParserException>(() => {
+                new ActionDecoderExtension().InvokeMatchingMethod(LINE_TO_PARSE);
+            });
+        }
+    }
+}
