+++ unity-ggjj/Assets/Tests/EditModeTests/TextDecoder/ActionDecoderTests/UnsupportedParameterTypeTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 418214285722bc74fbc3361a5fa42cab
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
